RAG/ARTEM_rag_template/requirements.txt
openai>=1.54.0
httpx>=0.27.0
PyYAML>=6.0
sentence-transformers>=3.1.0
numpy<2.0.0
rank-bm25>=0.2.2
deepeval>=1.0.0
tqdm>=4.66.0
faiss-cpu>=1.7.4
lightgbm>=4.3.0
catboost>=1.2.5
scikit-learn>=1.4.0
joblib>=1.3.0
pypdf>=5.0.0
transformers>=4.44.0
accelerate>=0.33.0
FlagEmbedding
einops
torch